use internal graph struct
