+++ internal/runner/options.go
@@ -1,15 +1,14 @@
 package runner
 
 import (
-	"github.com/dominikbraun/graph"
-
+	"github.com/BolajiOlajide/kat/internal/graph"
 	"github.com/BolajiOlajide/kat/internal/types"
 )
 
 // Options represents the options for the runner.
 type Options struct {
 	Operation     types.MigrationOperationType
-	Definitions   graph.Graph[int64, types.Definition]
+	Definitions   *graph.Graph
 	MigrationInfo types.MigrationInfo
 	DryRun        bool
 	Verbose       bool
